% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpubmed_textsearch.R
\name{abstract_to_text}
\alias{abstract_to_text}
\title{concatenates abstract list to a single string}
\usage{
abstract_to_text(article)
}
\arguments{
\item{article}{R representation of a pubmed record}
}
\description{
concatenates abstract list to a single string
}

