\name{chunker}
\alias{chunker}
\title{Helper function to split a vector v into list of chunks of chunk_size}
\usage{
chunker(v, chunk_size)
}
\arguments{
  \item{v}{a vector to be split into chunks}

  \item{chunk_size}{Integer size of the individual chunks}
}
\description{
Helper function to split a vector v into list of chunks of
chunk_size
}

