% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpubmed_locations.R
\name{geocode}
\alias{geocode}
\title{Helper function for geocode_address}
\usage{
geocode(address)
}
\arguments{
\item{address}{A string with an address}
}
\description{
Helper function for geocode_address
}

