\name{get_mesh_headings}
\alias{get_mesh_headings}
\title{Returns a list of MeSH headings for an article}
\usage{
get_mesh_headings(article)
}
\arguments{
  \item{article}{List representing a single Pubmed Record
  e.g. an element from a list returned from
  fetch_in_chunks()}
}
\value{
list of MeSH Headings
}
\description{
Returns a list of MeSH headings for an article
}

