\name{in_record_text_p}
\alias{in_record_text_p}
\title{predicate function for searching in title and abstract}
\usage{
in_record_text_p(article, terms, case_sensitive = FALSE, ...)
}
\arguments{
  \item{article}{R representation of a pubmed record}

  \item{terms}{vector of terms}

  \item{case_sensitive}{logical}

  \item{...}{arguments to be passed to grep}
}
\description{
predicate function for searching in title and abstract
}

