\name{mesh_heading_frequency}
\alias{mesh_heading_frequency}
\title{Returns a data frame of all MeSH headings in a corpus, with frequencies for each}
\usage{
mesh_heading_frequency(corpus)
}
\arguments{
  \item{article}{List representing a single Pubmed Record
  e.g. an element from a list returned from
  fetch_in_chunks()}
}
\value{
list of MeSHHeadings
}
\description{
Returns a data frame of all MeSH headings in a corpus, with
frequencies for each
}

