% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpubmed_mesh.R
\name{mesh_table}
\alias{mesh_table}
\title{helper function for mesh_assoc_table}
\usage{
mesh_table(articles)
}
\arguments{
\item{articles}{collection of pubmed articles}
}
\description{
helper function for mesh_assoc_table
}

