% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpubmed_textsearch.R
\name{mesh_to_text}
\alias{mesh_to_text}
\title{concatenates a list of MeSH headings to a single string}
\usage{
mesh_to_text(article)
}
\arguments{
\item{article}{R representation of a pubmed record}
}
\description{
concatenates a list of MeSH headings to a single string
}

