% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpubmed_io.R
\name{read_article_json}
\alias{read_article_json}
\title{Redundant wrapper around fromJSON}
\usage{
read_article_json(filename, ...)
}
\arguments{
\item{filename}{a filename to read from}

\item{...}{arguments to be passed to fromJSON}
}
\description{
Redundant wrapper around fromJSON
}

