\name{write_JSON_file}
\alias{write_JSON_file}
\title{Write a list of (e.g.) Pubmed records (e.g.) from rpubmed_fetch_in_chunks to json file}
\usage{
write_JSON_file(x, file)
}
\arguments{
  \item{x}{object to be serialised}

  \item{file}{filename}
}
\description{
Write a list of (e.g.) Pubmed records (e.g.) from
rpubmed_fetch_in_chunks to json file
}

