% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpubmed_textsearch.R
\name{in_abstract_p}
\alias{in_abstract_p}
\title{predicate function for searching abstracts}
\usage{
in_abstract_p(article, terms, case_sensitive = FALSE, ...)
}
\arguments{
\item{article}{R representation of a pubmed record}

\item{terms}{vector of terms}

\item{case_sensitive}{logical}

\item{...}{arguments to be passed to grep}
}
\description{
predicate function for searching abstracts
}

