\name{record_counts_by_year}
\alias{record_counts_by_year}
\title{Gives a breakdown of records per year in a corpus of Pubmed Records}
\usage{
  record_counts_by_year(corpus)
}
\arguments{
  \item{corpus}{a list of Pubmed records e.g. as given by
  fetch_in_chunks()}
}
\value{
  dataframe with year and Number_records columns
}
\description{
  Gives a breakdown of records per year in a corpus of
  Pubmed Records
}

