% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpubmed_textsearch.R
\name{term_in_text_p}
\alias{term_in_text_p}
\title{predicate function for presence of a term in an article text}
\usage{
term_in_text_p(term, text, case_sensitive, ...)
}
\arguments{
\item{term}{string}

\item{text}{article text}

\item{case_sensitive}{logical}

\item{...}{arguments to be passed to grep}
}
\description{
predicate function for presence of a term in an article text
}

